project_xfz/xfz/requirements.txt
aiohttp==3.3.2
amqp==1.4.9
anyjson==0.3.3
apptools==4.4.0
asn1crypto==0.24.0
astroid==2.0.4
async-timeout==3.0.0
attrs==18.1.0
Automat==0.7.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.4.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.3
cssselect==1.0.3
defusedxml==0.5.0
Django==2.0.7
django-celery==3.2.2
django-redis-sessions==0.6.1
django-tinymce==2.7.0
envisage==4.7.0
Flask==1.0.2
Flask-Login==0.4.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsmin==2.2.2
kombu==3.0.37
lazy-object-proxy==1.3.1
lxml==4.2.3
MarkupSafe==1.0
mayavi==4.6.2+vtk81
mccabe==0.6.1
msgpack==0.5.6
multidict==4.3.1
mysqlclient==1.4.2
numpy==1.15.0
pandas==0.23.4
parsel==1.5.0
Pillow==5.2.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pycurl==7.43.1
PyDispatcher==2.0.5
pyface==6.0.0
pygame==1.9.3
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==2.1.1
pymongo==3.7.1
PyMySQL==0.9.2
pyOpenSSL==18.0.0
PyQt4==4.11.4
pyquery==1.4.0
pyspider==0.3.10
python-dateutil==2.7.3
python-memcached==1.59
python-scrapyd-api==2.1.2
pytz==2018.5
pywin32==221
PyYAML==3.13
pyzmq==17.1.2
queuelib==1.5.0
redis==2.10.6
requests==2.19.1
Scrapy==1.5.1
scrapyd==1.2.0
selenium==3.13.0
service-identity==17.0.0
simplejson==3.16.0
six==1.11.0
tblib==1.3.2
tornado==5.1
traits==4.6.0
traitsui==6.0.0
tushare==1.2.12
Twisted==18.7.0
u-msgpack-python==2.5.0
urllib3==1.23
vine==1.1.4
virtualenv==16.4.0
virtualenvwrapper-win==1.2.5
VTK==8.1.1
w3lib==1.19.0
Werkzeug==0.14.1
wrapt==1.10.11
WsgiDAV==2.4.1
yarl==1.2.6
zope.interface==4.5.0
